requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
ccimport==0.4.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
filelock==3.12.4
fire==0.5.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.18.0
hydra-core==1.3.2
idna==3.4
lark==1.1.8
lightning==2.1.0
lightning-utilities==0.9.0
multidict==6.0.4
munch==4.0.0
ninja==1.11.1.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
open3d-python==0.3.0.0
packaging==23.2
pandas==2.1.1
pathlib2==2.3.7.post1
pathtools==0.1.2
pccm==0.4.9
Pillow==10.1.0
portalocker==2.8.2
pretrainedmodels==0.7.4
promise==2.3
protobuf==4.24.4
psutil==5.9.6
pybind11==2.11.1
python-dateutil==2.8.2
pytorch-lightning==2.1.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.0
scipy==1.11.3
segmentation-models-pytorch==0.3.3
sentry-sdk==1.32.0
setproctitle==1.3.3
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
termcolor==2.3.0
timm==0.9.2
torchmetrics==0.11.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wandb==0.13.5
yarl==1.9.2
